check if hopper is locked via redstone
