--- src/main/java/me/jellysquid/mods/lithium/common/blockentities/HopperCooldown.java
@@ -1,9 +1,0 @@
-package me.jellysquid.mods.lithium.common.blockentities;
-
-public interface HopperCooldown {
-	int cooldown();
-
-	default boolean cooled() {
-		return cooldown() <= 0;
-	}
-}
