--- src/main/java/me/jellysquid/mods/lithium/common/blockentities/HopperEntities.java
@@ -1,8 +1,0 @@
-package me.jellysquid.mods.lithium.common.blockentities;
-
-import net.minecraft.entity.ItemEntity;
-import java.util.List;
-
-public interface HopperEntities {
-	List<ItemEntity> entities();
-}
