+++ src/main/resources/lithium.mixins.json
@@ -37,9 +37,7 @@
     "small_tag_arrays.MixinTag",
     "voxelshape.fast_shape_vertex_merging.MixinVoxelShapes",
     "voxelshape.precompute_shape_arrays.MixinSimpleVoxelShape",
-    "entity.item_self_insert.standard.HopperBlockEntityMixin",
-    "entity.item_self_insert.standard.ItemEntityMixin",
-    "entity.item_self_insert.vanilish.HopperBlockEntityMixin",
-    "entity.item_self_insert.vanilish.ItemEntityMixin"
+    "entity.item_self_insert.HopperBlockEntityMixin",
+    "entity.item_self_insert.ItemEntityMixin"
   ]
 }
